🧹 Refactoring

+++ internal/service/user/user.go
@@ -40,7 +40,6 @@ func (s *serv) Create(ctx context.Context, user *model.UserCreate) (int64, error
 
 		return nil
 	})
-
 	if err != nil {
 		if errors.Is(err, ErrUserExists) {
 			return 0, ErrUserExists
@@ -69,7 +68,6 @@ func (s *serv) Get(ctx context.Context, id int64) (*model.User, error) {
 
 		return nil
 	})
-
 	if err != nil {
 		return nil, errors.New("failed to read user info")
 	}
@@ -79,6 +77,11 @@ func (s *serv) Get(ctx context.Context, id int64) (*model.User, error) {
 func (s *serv) Update(ctx context.Context, user *model.UserUpdate) error {
 	err := s.txManager.ReadCommitted(ctx, func(ctx context.Context) error {
 		var errTx error
+		_, errTx = s.userRepository.Get(ctx, user.ID)
+		if errTx != nil {
+			return errTx
+		}
+
 		errTx = s.userRepository.Update(ctx, user)
 		if errTx != nil {
 			return errTx
@@ -93,7 +96,6 @@ func (s *serv) Update(ctx context.Context, user *model.UserUpdate) error {
 
 		return nil
 	})
-
 	if err != nil {
 		return errors.New("failed to update user info")
 	}
@@ -103,6 +105,11 @@ func (s *serv) Update(ctx context.Context, user *model.UserUpdate) error {
 func (s *serv) Delete(ctx context.Context, id int64) error {
 	err := s.txManager.ReadCommitted(ctx, func(ctx context.Context) error {
 		var errTx error
+		_, errTx = s.userRepository.Get(ctx, id)
+		if errTx != nil {
+			return errTx
+		}
+
 		errTx = s.userRepository.Delete(ctx, id)
 		if errTx != nil {
 			return errTx
@@ -117,7 +124,6 @@ func (s *serv) Delete(ctx context.Context, id int64) error {
 
 		return nil
 	})
-
 	if err != nil {
 		return errors.New("failed to delete user")
 	}
